pyproject.toml
[project]
name = "test-aoc"
version = "0.1.0"
description = "Python project for solving Advent of Code challenges"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.1",
    "beautifulsoup4>=4.12.3",
    "httpx>=0.28.1",
    "markdownify>=0.14.1",
    "psycopg2-binary>=2.9.9",
    "typer>=0.15.1",
    "types-beautifulsoup4>=4.12.0",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D212"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "INP001"]
"dossier/mcp_server.py" = ["T201", "BLE001"]
"dossier/http_server.py" = ["T201", "BLE001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
